+++ tests/document/test_instance.py
@@ -46,6 +46,7 @@ from tests.utils import (
     MongoDBTestCase,
     db_ops_tracker,
     get_as_pymongo,
+    requires_mongodb_gte_44,
 )
 
 TEST_IMAGE_PATH = os.path.join(os.path.dirname(__file__), "../fields/mongoengine.png")
@@ -1875,6 +1876,53 @@ class TestDocumentInstance(MongoDBTestCase):
         person = self.Person.objects.get()
         assert not person.comments_dict["first_post"].published
 
+    @requires_mongodb_gte_44
+    def test_update_propagates_hint_collation_and_comment(self):
+        """Make sure adding a hint/comment/collation to the query gets added to the query"""
+        mongo_ver = get_mongodb_version()
+
+        base = {"locale": "en", "strength": 2}
+        index_name = "name_1"
+
+        class AggPerson(Document):
+            name = StringField()
+            meta = {
+                "indexes": [{"fields": ["name"], "name": index_name, "collation": base}]
+            }
+
+        AggPerson.drop_collection()
+        _ = AggPerson.objects.first()
+
+        comment = "test_comment"
+
+        if PYMONGO_VERSION >= (4, 1):
+            with db_ops_tracker() as q:
+                _ = AggPerson.objects.comment(comment).update_one(name="something")
+                query_op = q.db.system.profile.find(
+                    {"ns": "mongoenginetest.agg_person"}
+                )[0]
+                CMD_QUERY_KEY = "command" if mongo_ver >= MONGODB_36 else "query"
+                assert "hint" not in query_op[CMD_QUERY_KEY]
+                assert query_op[CMD_QUERY_KEY]["comment"] == comment
+                assert "collation" not in query_op[CMD_QUERY_KEY]
+
+        with db_ops_tracker() as q:
+            _ = AggPerson.objects.hint(index_name).update_one(name="something")
+            query_op = q.db.system.profile.find({"ns": "mongoenginetest.agg_person"})[0]
+            CMD_QUERY_KEY = "command" if mongo_ver >= MONGODB_36 else "query"
+
+            assert query_op[CMD_QUERY_KEY]["hint"] == {"$hint": index_name}
+            assert "comment" not in query_op[CMD_QUERY_KEY]
+            assert "collation" not in query_op[CMD_QUERY_KEY]
+
+        with db_ops_tracker() as q:
+            _ = AggPerson.objects.collation(base).update_one(name="something")
+            query_op = q.db.system.profile.find({"ns": "mongoenginetest.agg_person"})[0]
+            CMD_QUERY_KEY = "command" if mongo_ver >= MONGODB_36 else "query"
+            assert "hint" not in query_op[CMD_QUERY_KEY]
+            assert "comment" not in query_op[CMD_QUERY_KEY]
+            assert query_op[CMD_QUERY_KEY]["collation"] == base
+
     def test_delete(self):
         """Ensure that document may be deleted using the delete method."""
         person = self.Person(name="Test User", age=30)
@@ -1883,6 +1931,7 @@ class TestDocumentInstance(MongoDBTestCase):
         person.delete()
         assert self.Person.objects.count() == 0
 
+    @requires_mongodb_gte_44
     def test_delete_propagates_hint_collation_and_comment(self):
         """Make sure adding a hint/comment/collation to the query gets added to the query"""
         mongo_ver = get_mongodb_version()